chore: Refactoring time formatting.
