+++ lib/main.dart
@@ -6,7 +6,6 @@ import 'package:todo/bloc/todo_bloc.dart';
 import 'package:todo/breakpoints.dart';
 import 'package:todo/stopwatch.dart';
 import 'package:todo/models/item.dart';
-import 'package:todo/utils.dart';
 
 // Keys used for testing
 final textfieldKey = UniqueKey();
@@ -326,6 +325,16 @@ class _ItemCardState extends State<ItemCard> {
   void dispose() {
     _timer.cancel();
     super.dispose();
+  }
+
+  // Formats milliseconds to human-readable time
+  // https://itnext.io/create-a-stopwatch-app-with-flutter-f0dc6a176b8a
+  String formatTime(int milliseconds) {
+    var secs = milliseconds ~/ 1000;
+    var hours = (secs ~/ 3600).toString().padLeft(2, '0');
+    var minutes = ((secs % 3600) ~/ 60).toString().padLeft(2, '0');
+    var seconds = (secs % 60).toString().padLeft(2, '0');
+    return "$hours:$minutes:$seconds";
   }
 
   // Start and stop timer button handler
