--- lib/utils.dart
@@ -1,9 +1,0 @@
-/// Formats milliseconds to human-readable time
-/// https://itnext.io/create-a-stopwatch-app-with-flutter-f0dc6a176b8a
-String formatTime(int milliseconds) {
-  var secs = milliseconds ~/ 1000;
-  var hours = (secs ~/ 3600).toString().padLeft(2, '0');
-  var minutes = ((secs % 3600) ~/ 60).toString().padLeft(2, '0');
-  var seconds = (secs % 60).toString().padLeft(2, '0');
-  return "$hours:$minutes:$seconds";
-}
